some changes to ex2 and changes to checker.
we need to figure out how to handle the case where the last dot is eaten because currently the program avoids it.

we also need to figure out how to handle cases where mdp explores states that are not necessary indexed properly. (i don't fully understand this issue but i do know it exists when i test other initial states (with ghosts).